requirements.txt
masonite==2.2.15
waitress==1.3.1
gunicorn==19.9.0
psycopg2==2.8.3
masonite-cli==2.2.0
cleo==0.6.5
iroha==0.0.5.4
pyjwt==1.7.1
sendgrid==6.1.0
nested_lookup==0.2.18
protobuf3_to_dict==0.1.5
